notes/Azure_GPT4o.py
"""
This file is what I see in GPT4o Azure OpenAI Service website as sample code.
It does not work. I guess this is a mistake in Microsoft. I modified GPT4o code based on GPT4V code and it works. 
- Yiyun Chen 20240524.
"""



import os
from openai import AzureOpenAI
from azure.identity import DefaultAzureCredential, get_bearer_token_provider
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

endpoint = os.getenv("AZURE_OPENAI_ENDPOINT")
deployment = os.getenv("CHAT_COMPLETIONS_DEPLOYMENT_NAME")
# search_endpoint = os.environ["SEARCH_ENDPOINT"]
# search_index = os.environ["SEARCH_INDEX"]
print(endpoint)
      
token_provider = get_bearer_token_provider(DefaultAzureCredential(), "https://cognitiveservices.azure.com/.default")
      
client = AzureOpenAI(
    azure_endpoint=endpoint,
    azure_ad_token_provider=token_provider,
    api_version="2024-02-01",
)
      
completion = client.chat.completions.create(
    model=deployment,
    messages=[
        {
            "role": "user",
            "content": "Who is DRI?",
        },
        {
            "role": "assistant",
            "content": "DRI stands for Directly Responsible Individual of a service. Which service are you asking about?"
        },
        {
            "role": "user",
            "content": "Opinion mining service"
        }
    ],)
'''
    extra_body={
        "data_sources": [
            {
                "type": "azure_search",
                "parameters": {
                    "endpoint": search_endpoint,
                    "index_name": search_index,
                    "authentication": {
                        "type": "system_assigned_managed_identity"
                    }
                }
            }
        ]
    }
'''
    

      
print(completion.to_json())