requirements.txt
tqdm
torch
torchvision
matplotlib
numpy
requests
Pillow  # This is for PIL

wandb
einops
tiktoken
scipy

pynput

pyinstaller

moviepy
pygame

python-dotenv

openai 
azure-identity